move all browse fixtures into fixtures/browse
