+++ test/fixtures/browse/all.js
