+++ test/fixtures/browse/depended.js
