+++ test/fixtures/browse/index.js
@@ -1,0 +1,11 @@
+module.exports = {
+  all: require('./all'),
+  author: require('./author'),
+  depended: require('./depended'),
+  dependents: require('./dependents'),
+  keyword: require('./keyword'),
+  recentauthors: require('./recentauthors'),
+  star: require('./starred'),
+  updated: require('./updated'),
+  userstar: require('./userstar'),
+}
