+++ test/fixtures/browse/starred.js
