+++ test/fixtures/browse/userstar.js
