--- test/fixtures/browseData.js
@@ -1,17 +1,0 @@
-exports.all = require('./fake-browse/all');
-
-exports.author = require('./fake-browse/author');
-
-exports.depended = require('./fake-browse/depended');
-
-exports.dependents = require('./fake-browse/dependents');
-
-exports.keyword = require('./fake-browse/keyword');
-
-exports.recentauthors = require('./fake-browse/recentauthors');
-
-exports.star = require('./fake-browse/starred');
-
-exports.updated = require('./fake-browse/updated');
-
-exports.userstar = require('./fake-browse/userstar');
