+++ test/user/profile.js
@@ -9,7 +9,7 @@ var Code = require('code'),
     nock = require("nock");
 
 var server;
-var fakeBrowse = require('../fixtures/browseData');
+var fakeBrowse = require('../fixtures/browse/index.js');
 
 var username1 = 'fakeuser',
     username2 = 'fakeusercli';
